feat(solobase): implement embed × web flow
Lifted from solobase-cli's legacy_build, with export-assets converted
from a cargo subprocess to a direct solobase_browser library call. Static
HTTP server extracted to cli::helpers::http_server for reuse.
