+++ crates/solobase/src/cli/flows/embed_web.rs
@@ -1,13 +1,54 @@
-//! Stub for embed × web flow. Implemented in Task 11.
+//! Embed × web: today's solobase-cli flow, with export-assets as a direct
+//! library call instead of a cargo subprocess.
 
 use std::path::Path;
+use std::process::Command;
 
 use anyhow::Result;
 
-pub async fn build(_repo_root: &Path, _release: bool) -> Result<()> {
-    anyhow::bail!("embed × web build: not implemented (Task 11)")
+use crate::cli::cmd;
+use crate::cli::helpers::{blocks, http_server, overlays};
+use crate::cli::legacy_config;
+
+pub async fn build(repo_root: &Path, release: bool) -> Result<()> {
+    let (cfg, _) = legacy_config::find_and_load(repo_root)?;
+
+    // 1. wafer build per block.
+    blocks::build_all(repo_root)?;
+
+    // 2. wasm-pack the user's cdylib.
+    let mut wp = Command::new("wasm-pack");
+    let mut args = vec!["build".to_string(), "--target".into(), "web".into()];
+    args.push(if release { "--release".into() } else { "--dev".into() });
+    args.push("--out-dir".into());
+    args.push(cfg.wasm.out_dir.clone());
+    wp.args(&args).current_dir(repo_root);
+    cmd::run("wasm-pack build", wp)?;
+
+    // 3. Bundle: write static assets + content-hash + render templates.
+    let dist_dir = repo_root.join(&cfg.wasm.out_dir);
+    let app = solobase_browser::tools::bundle::AppConfig {
+        app_name: Some(cfg.app.name.clone()),
+        app_title: Some(cfg.app.title.clone()),
+        boot_redirect: Some(cfg.app.boot_redirect.clone()),
+        extra_bypass_prefix: cfg.assets.extra_bypass_prefix.clone(),
+    };
+    solobase_browser::assets::write_to(&dist_dir)?;
+    solobase_browser::tools::bundle::run(&dist_dir, repo_root, !release, app)?;
+
+    // 4. Apply overlays.
+    overlays::apply_overlays(&cfg, repo_root, &dist_dir)?;
+
+    let profile = if release { "release" } else { "dev" };
+    println!("built {} ({profile}) → {}", cfg.app.name, cfg.wasm.out_dir);
+    Ok(())
 }
 
-pub async fn serve(_repo_root: &Path, _release: bool, _port: Option<u16>) -> Result<()> {
-    anyhow::bail!("embed × web serve: not implemented (Task 11)")
+pub async fn serve(repo_root: &Path, release: bool, port: Option<u16>) -> Result<()> {
+    build(repo_root, release).await?;
+    let port = port.unwrap_or(8080);
+    let cfg = legacy_config::find_and_load(repo_root)?.0;
+    let dist = repo_root.join(&cfg.wasm.out_dir);
+    eprintln!("serving {} on http://127.0.0.1:{port}", dist.display());
+    http_server::serve_static(&dist, port).await
 }
