+++ crates/solobase/src/cli/helpers/mod.rs
@@ -1,4 +1,5 @@
 pub mod blocks;
 pub mod frontend;
+pub mod http_server;
 pub mod overlays;
 pub mod wasm;
